chore: refactor bloc example
