+++ examples/bloc/lib/di/providers.dart
@@ -1,0 +1,4 @@
+import 'package:bloc_example/cubit/theme_cubit.dart';
+import 'package:disco/disco.dart';
+
+final themeProvider = Provider((_) => ThemeCubit());
