+++ examples/bloc/lib/ui/screens/theme_switcher_page.dart
@@ -1,0 +1,16 @@
+import 'package:bloc_example/ui/widgets/themed_button.dart';
+import 'package:flutter/material.dart';
+
+class ThemeSwitcherPage extends StatelessWidget {
+  const ThemeSwitcherPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Cubit Theme Example')),
+      body: const Center(
+        child: ThemedButton(),
+      ),
+    );
+  }
+}
