+++ examples/bloc/pubspec.yaml
@@ -9,7 +9,8 @@ environment:
 resolution: workspace
 
 dependencies:
-  disco: ^0.0.2
+  disco:
+    path: ../../packages/disco
   flutter:
     sdk: flutter
   flutter_bloc: ^9.0.0
